First Commit
This is the first commit. It's simply a starting point for the maze
program. We will need to add a JFrame for our main starting point with a
nice graphic or title centered on the frame and some buttons at the
bottom to give the user the option of creating their own maze our to use
a random maze that we already generated... one though on the user
generated maze, we could give them the option of building an easy,
medium, or hard maze. then we could populate the board with more or less
cells.